Change package names and implement first communication
Raspberry pi is now able to establish a connection to the Server.
The Server will save the Raspberry Pi in its managed Pis and save the Remote Interface for configuration in its inteface list.
Upon shutdown, the Raspberry pi will send a disconnect to the Server, which will remove the remote interface and send an OK message back to the raspberry pi